Consolidate Programs and Community into single Community page
The Programs page content (free activities, continued programs, activity
cards) is merged into the Community page, which now serves as the single
destination for community vision and ongoing programs. The /programs
route redirects to /community, and the Programs nav item is removed
to reduce navigation clutter.
